requirements.txt
# Основні залежності для Raspberry Pi 5

# GPIO та апаратне забезпечення
RPi.GPIO==0.7.1

# Робота з камерами
picamera2==0.3.12

# Обробка зображень
opencv-python==4.8.1.78
numpy==1.24.3

# Машинне навчання
onnxruntime==1.16.3
torch==2.1.0
torchvision==0.16.0

# Google Sheets API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Допоміжні бібліотеки
Pillow==10.1.0
scipy==1.11.4